LPS-104616 Move updateLayoutData to thunks
